Post and Get methods working, Delete note function still not working correctly.
