+++ db/db.json
@@ -1,1 +1,1 @@
-[object Object]
+[{"title":"Second Note","text":"Hey","id":19471},{"title":"note","text":"hey","id":82131}]
