+++ db/journal.json
@@ -1,0 +1,17 @@
+[
+    {
+        "title": "tests",
+        "body": "asfdfsd",
+        "journal": "food"
+    },
+    {
+        "title": "lskfjsdlk",
+        "body": "sdjlfkjdsklf",
+        "journal": "excercise"
+    },
+    {
+        "title": "ouch",
+        "body": "dude",
+        "journal": "food"
+    }
+]
