+++ package.json
@@ -10,6 +10,8 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.16.4",
-    "mysql": "^2.16.0"
+    "file-system": "^2.2.2",
+    "mysql": "^2.16.0",
+    "path": "^0.12.7"
   }
 }
