+++ public/assets/js/index.js
@@ -47,6 +47,7 @@ var renderActiveNote = function() {
     $noteTitle.val("");
     $noteText.val("");
   }
+  location.reload(true);
 };
 
 // Get the note data from the inputs, save it to the db and update the view
@@ -59,6 +60,7 @@ var handleNoteSave = function() {
   saveNote(newNote).then(function(data) {
     getAndRenderNotes();
     renderActiveNote();
+    location.reload();
   });
 };
