--- public/manifest.webmanifest
@@ -1,8 +1,0 @@
-{
-    "name": "Notetaker App",
-    "short_name": "NT App",
-    "start_url": "/index.html",
-    "theme_color": "#ffffff",
-    "background_color": "#ffffff",
-    "display": "standalone"
-}
