+++ routes/htmlRoutes.js
@@ -1,0 +1,12 @@
+var path = require("path");
+//Basic Get Functions
+module.exports = function (app) {
+    //HTML routes
+    app.get("/notes", function (req, res) {
+        res.sendFile(path.join(__dirname, "../public/notes.html"));
+    });
+
+    app.get("/", function (req, res) {
+        res.sendFile(path.join(__dirname, "../public/index.html"));
+    });
+}
